+++ types/src/schemas/query.rs
@@ -1,9 +1,7 @@
 use serde::Deserialize;
 
-
-
 #[derive(Debug, Deserialize)]
-pub struct ListAmount{
-    pub offset: Option<usize>,
-    pub max: Option<usize>
+pub struct ListAmount {
+    pub offset: Option<i64>,
+    pub max: Option<i64>,
 }
